Extract mock users into their own test data file
