+++ server/routes/testutils/appSetup.ts
@@ -10,50 +10,11 @@ import nunjucksSetup from '../../utils/nunjucksSetup'
 import type { ApplicationInfo } from '../../applicationInfo'
 import errorHandler from '../../errorHandler'
 import type { Services } from '../../services'
-import { makeMockCaseload, mockCaseload } from '../../data/testData/frontendComponents'
-import { leeds } from '../../data/testData/prisonApi'
-import { roleReadOnly, roleReadWrite, roleApproveReject } from '../../data/constants'
+import { reportingUser } from '../../data/testData/users'
 import { IncidentReportingApi } from '../../data/incidentReportingApi'
 import { OffenderSearchApi } from '../../data/offenderSearchApi'
 import { PrisonApi } from '../../data/prisonApi'
 import { setupPermissions } from '../../middleware/permissions'
-
-/** Typical reporting officer with access to Moorland only */
-export const user: Express.User = {
-  name: 'JOHN SMITH',
-  userId: 'id',
-  token: 'token',
-  username: 'user1',
-  displayName: 'John Smith',
-  active: true,
-  activeCaseLoadId: 'MDI',
-  authSource: 'NOMIS',
-  roles: ['PRISON', roleReadWrite],
-  // from frontend components middleware
-  activeCaseLoad: mockCaseload,
-  caseLoads: [mockCaseload],
-}
-
-/** Data warden with write access to Leeds and Moorland */
-export const approverUser: Express.User = {
-  ...user,
-  roles: ['PRISON', roleApproveReject],
-  caseLoads: [mockCaseload, makeMockCaseload(leeds, false)],
-}
-
-/** HQ user with read-only access to Leeds and Moorland */
-export const hqUser: Express.User = {
-  ...user,
-  roles: ['PRISON', roleReadOnly],
-  caseLoads: [mockCaseload, makeMockCaseload(leeds, false)],
-}
-
-/** General user in Moorland without access */
-export const unauthorisedUser: Express.User = {
-  ...user,
-  roles: ['PRISON'],
-  caseLoads: [mockCaseload],
-}
 
 export const testAppInfo: ApplicationInfo = {
   applicationName: 'hmpps-incident-reporting',
@@ -103,7 +64,7 @@ function appSetup(services: Services, production: boolean, userSupplier: () => E
 export function appWithAllRoutes({
   production = false,
   services = { applicationInfo: testAppInfo },
-  userSupplier = () => user,
+  userSupplier = () => reportingUser,
 }: {
   production?: boolean
   services?: Partial<Services>
